refactor: Organization 타입 분리
