+++ client/src/features/Organization/Select/types/Organization.ts
@@ -1,0 +1,10 @@
+export type Organization = {
+  organizationId: number;
+  name: string;
+  description: string;
+  imageUrl: string;
+};
+
+export type OrganizationWithRole = Organization & {
+  isAdmin: boolean;
+};
